requirements.txt
bokeh
paho-mqtt
tensorflow
scikit-learn
pandas
numpy
